+++ example/lib/main.dart
@@ -35,23 +35,24 @@ class MyApp extends StatelessWidget {
                       '''Who else thinks it's thinks it's just cool to mention 
                       @jane when #JaneMustLive is trending without even trying 
                       to send a *bold* email to janedoe@gmail.com and verify the
-                       facts talkmore of visiting www.janedoe.com''',
+                       facts talkmore of ivisitingi www.janedoe.com''',
                   maxLines: 3,
-                  onEmailClicked: (email) => print('$email clicked'),
-                  onHashTagClicked: (hashtag) => print('is $hashtag trending?'),
-                  onMentionClicked: (mention) => print('$mention clicked'),
-                  onUrlClicked: (url) => print('visting $url?'),
-                  onPhoneClicked: (phone) => print('click phone $phone'),
                   truncate: true,
                   viewLessText: 'less',
                   linkStyle: TextStyle(color: Colors.blue),
+                  textAlign: TextAlign.center,
                   supportedTypes: [
-                    ParsedType.EMAIL,
-                    ParsedType.HASH,
-                    ParsedType.MENTION,
-                    ParsedType.URL,
-                    ParsedType.BOLD,
-                    ParsedType.PHONE,
+                    EmailParser(
+                        onTap: (email) => print('${email.value} clicked')),
+                    PhoneParser(
+                        onTap: (phone) => print('click phone ${phone.value}')),
+                    MentionParser(
+                        onTap: (mention) => print('${mention.value} clicked')),
+                    UrlParser(onTap: (url) => print('visting ${url.value}?')),
+                    BoldParser(),
+                    HashTagParser(
+                        onTap: (hashtag) =>
+                            print('is ${hashtag.value} trending?'))
                   ],
                 ),
               ),